+++ src/olmo_core/data/__init__.py
@@ -3,6 +3,7 @@ from .iterable_dataset import IterableDataset
 from .mixes import DataMix
 from .numpy_dataset import NumpyDataset, NumpyDatasetConfig, NumpyDatasetDType
 from .tokenizer import TokenizerConfig, TokenizerName
+from .utils import melt_batch, split_batch, truncate_batch
 
 __all__ = [
     "NumpyDatasetConfig",
@@ -13,5 +14,8 @@ __all__ = [
     "DataMix",
     "DataCollator",
     "PaddingDirection",
+    "split_batch",
+    "melt_batch",
+    "truncate_batch",
     "IterableDataset",
 ]