+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added support for sequence length warm-up.
+- Added support for changing train sequence length when loading a checkpoint.
+- Added support for sequence length warm-up during training via the callback `SequenceLengthSchedulerCallback`.
 - Added `Lion` and `SkipStepLion` optimizers.
 
 ### Changed